requirements.txt
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
dnspython>=2.3.0
python-whois>=0.8.0
lxml>=4.9.0
colorama>=0.4.6
pystyle>=2.9
requests>=2.28.0
phonenumbers>=8.13.0
piexif>=1.1.3
exifread>=3.0.0
Pillow>=10.0.0
prettytable>=3.0.0
